async rerank & add searchuploaddocs class
